// web/src/api/generated/models/TokenView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TokenView
 */
export interface TokenView {
    /**
     * 
     * @type {Date}
     * @memberof TokenView
     */
    accessExpiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TokenView
     */
    accessToken: string;
    /**
     * 
     * @type {Date}
     * @memberof TokenView
     */
    refreshExpiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TokenView
     */
    refreshToken: string;
}

/**
 * Check if a given object implements the TokenView interface.
 */
export function instanceOfTokenView(value: object): value is TokenView {
    if (!('accessExpiresAt' in value) || value['accessExpiresAt'] === undefined) return false;
    if (!('accessToken' in value) || value['accessToken'] === undefined) return false;
    if (!('refreshExpiresAt' in value) || value['refreshExpiresAt'] === undefined) return false;
    if (!('refreshToken' in value) || value['refreshToken'] === undefined) return false;
    return true;
}

export function TokenViewFromJSON(json: any): TokenView {
    return TokenViewFromJSONTyped(json, false);
}

export function TokenViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): TokenView {
    if (json == null) {
        return json;
    }
    return {
        
        'accessExpiresAt': (new Date(json['accessExpiresAt'])),
        'accessToken': json['accessToken'],
        'refreshExpiresAt': (new Date(json['refreshExpiresAt'])),
        'refreshToken': json['refreshToken'],
    };
}

export function TokenViewToJSON(json: any): TokenView {
    return TokenViewToJSONTyped(json, false);
}

export function TokenViewToJSONTyped(value?: TokenView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'accessExpiresAt': value['accessExpiresAt'].toISOString(),
        'accessToken': value['accessToken'],
        'refreshExpiresAt': value['refreshExpiresAt'].toISOString(),
        'refreshToken': value['refreshToken'],
    };
}